overlay: Shorten overly long Wi-Fi calling operator names
Some of the Wi-Fi calling operator names are too long to be rendered in
the top-right corner of the quick settings panel when they're formatted
with a carrier name, e.g.:

    T-Mobile Wi-Fi Calling
    WiFi Calling | T-Mobile

These overly long names cause the system to fall back to showing them
with a marquee in the expanded quick settings panel, which is ugly and
causes continuous frame rendering while the panel is expanded for no
good reason.

The longer names don't really provide any extra value anyway, so let's
just replace both of them with "%s Wi-Fi" to shorten them such that they
can fit in the corner of the QS panel.

Change-Id: Ie60561905201f7a88f01993e0ad4aca53b35c95a